tcp: Clear buffer on new connection
While TCP ensures that individual packets arrive whole and in order,
there is no guarantee that a frame arrives in a single packet,
which means some of it could be lost.

Instead of retransmitting the rest of the frame (as in that time
the animation pattern will change anyway, so it's better to just
send a new frame), we will abandon the stream on any connection
issues and restart transmitting with the next entire frame.

In order for that to work, receiver needs to drop any partial
frames it received in the meantime, so that the beginning of the
next frame is not interpreted as the end of the previous one,
and some arbitrary bytes in the middle of it aren't treated
as the length of the next one.

Additionally, the connection is abandoned much faster in case
no data is received, to allow for a faster reconnection.
For that reason the timeout was lowered to 1 second, and keepalive
packets were turned off entirely.